functions.js changed
