+++ js/comment/comment-util.js
@@ -1,4 +1,4 @@
-import { getRandomInteger, getRandomArrayElement } from './util.js';
+import { getRandomInteger, getRandomArrayElement } from '../util.js';
 
 const MESSAGE = [
   'Всё отлично!',
